+++ packages/create-electronup/src/types/degit.d.ts
@@ -1,0 +1,4 @@
+declare module 'degit' {
+  const degit: any
+  export default degit
+}